+++ examples/kitchen-sink/src/kitchen-sink-definitions.ts
@@ -6,7 +6,18 @@ import {
   arg,
   extendType,
   scalarType,
+  extendInputType,
+  intArg,
+  idArg,
 } from "nexus";
+
+export const testArgs1 = {
+  foo: idArg(),
+};
+
+export const testArgs2 = {
+  bar: idArg(),
+};
 
 export const Bar = interfaceType({
   name: "Bar",
@@ -22,6 +33,9 @@ export const Bar = interfaceType({
             answer: 2,
           },
         }),
+      },
+      resolve(root, args) {
+        return true;
       },
     });
     t.resolveType((root) => "Foo");
@@ -72,12 +86,27 @@ export const InputType = inputObjectType({
   },
 });
 
+export const ext = extendInputType({
+  type: "InputType",
+  definition(t) {},
+});
+
 export const Query = objectType({
   name: "Query",
   definition(t) {
     t.field("bar", {
       type: "Bar",
       resolve: () => ({ ok: true }),
+    });
+    t.int("getNumberOrNull", {
+      nullable: true,
+      args: { a: intArg({ required: true }) },
+      async resolve(_, { a }) {
+        if (a > 0) {
+          return a;
+        }
+        return null;
+      },
     });
   },
 });
@@ -87,6 +116,9 @@ export const MoreQueryFields = extendType({
   definition(t) {
     t.field("extended", {
       type: "Bar",
+      resolve() {
+        return { ok: true };
+      },
     });
   },
 });