+++ src/builder.ts
@@ -32,6 +32,7 @@ import {
   isOutputType,
   isUnionType,
   isScalarType,
+  defaultFieldResolver,
 } from "graphql";
 import { NexusArgConfig, NexusArgDef } from "./definitions/args";
 import {
@@ -76,17 +77,27 @@ import {
   isNexusUnionTypeDef,
   isNexusWrappedType,
   NexusWrappedType,
+  isNexusExtendInputTypeDef,
 } from "./definitions/wrapping";
 import {
   GraphQLPossibleInputs,
   GraphQLPossibleOutputs,
   NonNullConfig,
+  WrappedResolver,
 } from "./definitions/_types";
 import { TypegenAutoConfigOptions } from "./typegenAutoConfig";
 import { TypegenFormatFn } from "./typegenFormatPrettier";
 import { TypegenMetadata } from "./typegenMetadata";
-import { AbstractTypeResolver, GetGen } from "./typegenTypeHelpers";
+import {
+  AbstractTypeResolver,
+  GetGen,
+  AuthorizeResolver,
+} from "./typegenTypeHelpers";
 import { firstDefined, objValues, suggestionList, isObject } from "./utils";
+import {
+  NexusExtendInputTypeDef,
+  NexusExtendInputTypeConfig,
+} from "./definitions/extendInputType";
 
 export type Maybe<T> = T | null;
 
@@ -118,7 +129,7 @@ export interface BuilderConfig {
     | false;
   /**
    * Whether the schema & types are generated when the server
-   * starts. Default is !process.env.NODE_ENV || process.env.NODE_ENV !== "development"
+   * starts. Default is !process.env.NODE_ENV || process.env.NODE_ENV === "development"
    */
   shouldGenerateArtifacts?: boolean;
   /**
@@ -215,6 +226,13 @@ export class SchemaBuilder {
     NexusExtendTypeConfig<string>[] | null
   > = {};
   /**
+   * All "extensions" to input types (adding fields on types from many locations)
+   */
+  protected inputTypeExtensionMap: Record<
+    string,
+    NexusExtendInputTypeConfig<string>[] | null
+  > = {};
+  /**
    * Configures the root-level nonNullDefaults defaults
    */
   protected nonNullDefaults: NonNullConfig = {};
@@ -240,6 +258,7 @@ export class SchemaBuilder {
   addType(
     typeDef:
       | AllNexusNamedTypeDefs
+      | NexusExtendInputTypeDef<string>
       | NexusExtendTypeDef<string>
       | GraphQLNamedType
   ) {
@@ -249,6 +268,13 @@ export class SchemaBuilder {
     if (isNexusExtendTypeDef(typeDef)) {
       const typeExtensions = (this.typeExtensionMap[typeDef.name] =
         this.typeExtensionMap[typeDef.name] || []);
+      typeExtensions.push(typeDef.value);
+      return;
+    }
+
+    if (isNexusExtendInputTypeDef(typeDef)) {
+      const typeExtensions = (this.inputTypeExtensionMap[typeDef.name] =
+        this.inputTypeExtensionMap[typeDef.name] || []);
       typeExtensions.push(typeDef.value);
       return;
     }
@@ -312,6 +338,15 @@ export class SchemaBuilder {
         });
       }
     });
+    Object.keys(this.inputTypeExtensionMap).forEach((key) => {
+      // If we haven't defined the type, assume it's an object type
+      if (this.inputTypeExtensionMap[key] !== null) {
+        this.buildInputObjectType({
+          name: key,
+          definition() {},
+        });
+      }
+    });
     return {
       typeMap: this.finalTypeMap,
     };
@@ -325,6 +360,13 @@ export class SchemaBuilder {
       addField: (field) => fields.push(field),
     });
     config.definition(this.withScalarMethods(definitionBlock));
+    const extensions = this.inputTypeExtensionMap[config.name];
+    if (extensions) {
+      extensions.forEach((extension) => {
+        extension.definition(definitionBlock);
+      });
+    }
+    this.inputTypeExtensionMap[config.name] = null;
     return this.finalize(
       new GraphQLInputObjectType({
         name: config.name,
@@ -506,9 +548,9 @@ export class SchemaBuilder {
     return type;
   }
 
-  protected withScalarMethods<T extends NexusGenCustomDefinitionMethods<string>>(
-    definitionBlock: T
-  ): T {
+  protected withScalarMethods<
+    T extends NexusGenCustomDefinitionMethods<string>
+  >(definitionBlock: T): T {
     this.customScalarMethods.forEach(([methodName, typeName]) => {
       // @ts-ignore - Yeah, yeah... we know
       definitionBlock[methodName] = function(fieldName, ...opts) {
@@ -888,6 +930,12 @@ export class SchemaBuilder {
     if (!resolver && !forInterface) {
       resolver = (typeConfig as NexusObjectTypeConfig<any>).defaultResolver;
     }
+    if (fieldOptions.authorize) {
+      resolver = wrapAuthorize(
+        resolver || defaultFieldResolver,
+        fieldOptions.authorize
+      );
+    }
     return resolver;
   }
 
@@ -907,6 +955,33 @@ function extendError(name: string) {
   return new Error(
     `${name} was already defined and imported as a type, check the docs for extending types`
   );
+}
+
+export function wrapAuthorize(
+  resolver: GraphQLFieldResolver<any, any>,
+  authorize: AuthorizeResolver<string, any>
+): GraphQLFieldResolver<any, any> {
+  const nexusAuthWrapped: WrappedResolver = async (root, args, ctx, info) => {
+    const authResult = await authorize(root, args, ctx, info);
+    if (authResult === true) {
+      return resolver(root, args, ctx, info);
+    }
+    if (authResult === false) {
+      throw new Error("Not authorized");
+    }
+    if (authResult instanceof Error) {
+      throw authResult;
+    }
+    const {
+      fieldName,
+      parentType: { name: parentTypeName },
+    } = info;
+    throw new Error(
+      `Nexus authorize for ${parentTypeName}.${fieldName} Expected a boolean or Error, saw ${authResult}`
+    );
+  };
+  nexusAuthWrapped.nexusWrappedResolver = resolver;
+  return nexusAuthWrapped;
 }
 
 export interface BuildTypes<
@@ -943,6 +1018,7 @@ function addTypes(builder: SchemaBuilder, types: any) {
   if (
     isNexusNamedTypeDef(types) ||
     isNexusExtendTypeDef(types) ||
+    isNexusExtendInputTypeDef(types) ||
     isNamedType(types)
   ) {
     builder.addType(types);