+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+### 0.9.17
+
+- More for #55, type error on Promise/null resolve
+
+### 0.9.16
+
+- Fix #52, args being incorrectly imported as types
+- Feature: extendInputType, #51
+
+### 0.9.15
+
+- Fix #55, type error for Promise numeric field return
+
 ### 0.9.14
 
 - Fix #44, add correct typings for union type